Refactor avalonia port
