+++ src/TomsToolbox.Avalonia/DispatcherExtensions.cs
@@ -1,0 +1,36 @@
+﻿#pragma warning disable IDE0130
+// ReSharper disable once CheckNamespace : Keep it namespace-compatible with the WPF pendant.
+namespace TomsToolbox.Wpf;
+
+/// <summary>
+/// Extensions for the <see cref="Dispatcher"/> class.
+/// </summary>
+public static class DispatcherExtensions
+{
+    /// <summary>Gets the <see cref="T:Dispatcher" /> for the UI thread and creates a new <see cref="T:Dispatcher" /> if one is not already associated with the thread.</summary>
+    /// <returns>The dispatcher associated with the UI thread.</returns>
+    public static Dispatcher CurrentDispatcher => Dispatcher.UIThread;
+
+    /// <summary>
+    /// Executes the specified method on the dispatcher thread asynchronously, using the WPF-style naming convention.
+    /// </summary>
+    /// <param name="dispatcher">The dispatcher.</param>
+    /// <param name="method">The method.</param>
+    /// <param name="priority">The priority</param>
+    public static void BeginInvoke(this Dispatcher dispatcher, Action method, DispatcherPriority priority = default)
+    {
+        dispatcher.Post(method, priority);
+    }
+
+    /// <summary>
+    /// Executes the specified action on the dispatcher thread asynchronously, using the WPF-style naming convention.
+    /// </summary>
+    /// <param name="dispatcher">The dispatcher.</param>
+    /// <param name="method">The action.</param>
+    /// <param name="priority">The priority</param>
+    /// <returns>A task representing the asynchronous operation.</returns>
+    public static void BeginInvoke(this Dispatcher dispatcher, DispatcherPriority priority, Action method)
+    {
+        dispatcher.BeginInvoke(method, priority);
+    }
+}
