+++ src/TomsToolbox.Avalonia/FodyWeavers.xml
@@ -1,5 +1,3 @@
 ﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <WeakEventHandler />
-  <Throttle />
   <JetBrainsAnnotations />
 </Weavers>
