+++ src/TomsToolbox.Avalonia/Properties/AssemblyInfo.cs
@@ -1,11 +1,8 @@
 ﻿using System.Reflection;
 using System.Resources;
 using System.Runtime.InteropServices;
-using System.Windows;
-using System.Windows.Markup;
 
 [assembly: AssemblyCulture("")]
-[assembly: CLSCompliant(true)]
+[assembly: NeutralResourcesLanguage("en-US")]
+[assembly: CLSCompliant(false)]
 [assembly: ComVisible(false)]
-
-[assembly: NeutralResourcesLanguage("en-US", UltimateResourceFallbackLocation.MainAssembly)]
