+++ src/TomsToolbox.Avalonia/TomsToolbox.Avalonia.csproj
@@ -16,9 +16,8 @@
 
   <ItemGroup>
     <Compile Include="..\TomsToolbox.Wpf\DispatcherThrottle.cs" Link="DispatcherThrottle.cs" />
-    <Compile Include="..\TomsToolbox.Wpf\ObservableObject.cs" Link="ObservableObject.cs" />
+    <Compile Include="..\TomsToolbox.Wpf\ObservableObjectBase.cs" Link="ObservableObjectBase.cs" />
     <Compile Include="..\TomsToolbox.Wpf\PropertyDependencyAttribute.cs" Link="PropertyDependencyAttribute.cs" />
-    <Compile Include="..\TomsToolbox.Wpf\RelayedEventAttribute.cs" Link="RelayedEventAttribute.cs" />
     <Compile Include="..\TomsToolbox.Wpf\Throttle.cs" Link="Throttle.cs" />
   </ItemGroup>
 
@@ -32,8 +31,6 @@
     <PackageReference Include="JetBrains.Annotations" Version="2024.3.0" PrivateAssets="All" />
     <PackageReference Include="JetBrainsAnnotations.Fody" Version="2.17.0" PrivateAssets="All" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
-    <PackageReference Include="Throttle.Fody" Version="1.7.0" PrivateAssets="All" />
-    <PackageReference Include="WeakEventHandler.Fody" Version="1.8.0" PrivateAssets="All" />
   </ItemGroup>
 
 </Project>
