+++ src/TomsToolbox.Wpf.Composition/Properties/AssemblyInfo.cs
@@ -5,7 +5,7 @@ using System.Windows;
 using System.Windows.Markup;
 
 [assembly: AssemblyCulture("")]
-[assembly: CLSCompliant(true)]
+[assembly: CLSCompliant(false)]
 [assembly: ComVisible(false)]
 
 [assembly: NeutralResourcesLanguage("en-US", UltimateResourceFallbackLocation.MainAssembly)]
