+++ src/TomsToolbox.Wpf/GlobalUsings.cs
@@ -1,0 +1,3 @@
+﻿global using Dispatcher = System.Windows.Threading.Dispatcher;
+global using DispatcherTimer = System.Windows.Threading.DispatcherTimer;
+global using DispatcherPriority = System.Windows.Threading.DispatcherPriority;
